requirements.txt
# Libraries needed for API calls, JSON handling, and secure key loading
requests
openai
python-dotenv